requirements.txt
django
djangorestframework
django-extensions
Werkzeug
coverage
